// microservicios-web/src/app/views/dashboard/Setup/dstienda/models/dstienda.ts
export class Dstienda {
  id?: number;
  nombre?: string;
  direccion?: string;
  ciudad?: string;
  telefono?: string;
  correoElectronico?: string;
  horarioApertura?: string;
  horarioCierre?: string;
  marcasDisponibles?: string;
  tiposProductos?: string;
  inventario?: string;
  ingresosMensuales?: string;
  sitioWeb?: string;
  redesSociales?: string;
  serviciosAdicionales?: string;
  promociones?: string;
}